requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
astroid==2.11.2
    # via pylint
bandit==1.7.4
    # via -r requirements.in
black==22.1.0
    # via -r requirements.in
certifi==2021.10.8
    # via requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via
    #   black
    #   pip-tools
    #   safety
configurator[yaml]==2.6.0
    # via -r requirements.in
dill==0.3.4
    # via pylint
distlib==0.3.4
    # via virtualenv
dparse==0.5.1
    # via safety
filelock==3.6.0
    # via virtualenv
flake8==4.0.1
    # via -r requirements.in
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via bandit
identify==2.4.12
    # via pre-commit
idna==3.3
    # via requests
isort==5.10.1
    # via
    #   -r requirements.in
    #   pylint
lazy-object-proxy==1.7.1
    # via astroid
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
packaging==21.3
    # via
    #   dparse
    #   safety
pathspec==0.9.0
    # via black
pbr==5.8.1
    # via stevedore
pep517==0.12.0
    # via pip-tools
pip-tools==6.5.1
    # via -r requirements.in
platformdirs==2.5.1
    # via
    #   black
    #   pylint
    #   virtualenv
pre-commit==2.17.0
    # via -r requirements.in
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pylint==2.13.1
    # via -r requirements.in
pyparsing==3.0.7
    # via packaging
pyyaml==6.0
    # via
    #   bandit
    #   configurator
    #   dparse
    #   pre-commit
requests==2.27.1
    # via safety
safety==1.10.3
    # via -r requirements.in
six==1.16.0
    # via virtualenv
smmap==5.0.0
    # via gitdb
stevedore==3.5.0
    # via bandit
toml==0.10.2
    # via
    #   dparse
    #   pre-commit
tomli==2.0.1
    # via
    #   black
    #   pep517
    #   pylint
typing-extensions==4.1.1
    # via
    #   astroid
    #   black
    #   pylint
urllib3==1.26.9
    # via requests
virtualenv==20.14.0
    # via pre-commit
wheel==0.37.1
    # via pip-tools
wrapt==1.14.0
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
